sparated search Sidebar component
